+++ src/scame/core/06_scame_keys.el
@@ -190,5 +190,6 @@
 (global-set-key (kbd "C-c s z") 'hydra-scame-zoom/body)
 
 
+
 (provide '06_scame_keys)
 ;;; 06_scame_keys.el ends here