+++ src/scame/gnus/gnus-local.el
@@ -1,6 +1,6 @@
 ;;; gnus-local.el --- Localhost IMAP configuration for Gnus
 
-;; Copyright (C) 2015 Nicolas Lamirault <nicolas.lamirault@gmail.com>
+;; Copyright (C) 2015, 2016 Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 ;; This program is free software: you can redistribute it and/or modify
 ;; it under the terms of the GNU General Public License as published by
@@ -33,6 +33,9 @@
 ;; - msmtp to send emails
 ;; ------------------------------------------
 
+;; Customization :
+;; (setq scame-maildir-local "/opt/Exchange")
+
 
 ;; IMAP Localhost offlineimap
 
@@ -43,11 +46,19 @@
 ;;                	 "/usr/lib/dovecot/imap -o mail_location=maildir:/opt/Exchange:LAYOUT=fs:INBOX=/opt/Exchange/INBOX")))
 
 ;; IMAP Localhost imapsync
+;; (setq gnus-select-method
+;;       '(nnimap "Exchange"
+;;                (nnimap-stream shell)
+;;                (nnimap-shell-program
+;;                	 "/usr/lib/dovecot/imap -o mail_location=maildir:/opt/Exchange:LAYOUT=fs:INBOX=/opt/Exchange/Inbox")))
+
 (setq gnus-select-method
-      '(nnimap "Exchange"
+      `(nnimap "Exchange"
                (nnimap-stream shell)
                (nnimap-shell-program
-               	 "/usr/lib/dovecot/imap -o mail_location=maildir:/opt/OrangeExchange:LAYOUT=fs:INBOX=/opt/OrangeExchange/Inbox")))
+                (concat "/usr/lib/dovecot/imap -o mail_location=maildir:"
+                        ,scame-maildir-local ":LAYOUT=fs:INBOX="
+                        ,scame-maildir-local "/Inbox"))))
 
 
 ;; (setq gnus-secondary-select-methods
@@ -81,7 +92,7 @@
 (defun lld-notmuch-file-to-group (file)
   "Calculate the Gnus group name from the given file name."
   (let ((group (file-name-directory (directory-file-name (file-name-directory file)))))
-    (setq group (replace-regexp-in-string ".*/opt/Exchange/" "" group))
+    (setq group (replace-regexp-in-string ".*/opt/OrangeExchange/" "" group))
     (setq group (replace-regexp-in-string "/$" "" group))
     (if (string-match ":$" group)
         (concat group "INBOX")