+++ src/scame/packages/01_basic_ui.el
@@ -1,6 +1,6 @@
 ;;; 01_basic_ui.el --- Emacs basic UI
 
-;; Copyright (c) 2014, 2015 Nicolas Lamirault <nicolas.lamirault@gmail.com>
+;; Copyright (c) 2014-2016 Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 ;; This program is free software: you can redistribute it and/or modify
 ;; it under the terms of the GNU General Public License as published by
@@ -38,7 +38,12 @@
   (use-package tool-bar
     :config (tool-bar-mode -1))
 
-  (setq x-select-enable-clipboard t))
+  (setq x-select-enable-clipboard t)
+
+  (use-package mode-icons
+    :config (mode-icons-mode))
+
+  )
 
 ;; Window moves
 (global-set-key (kbd "C-c <left>")  'windmove-left)
@@ -55,6 +60,19 @@
 (use-package ace-window
   :config (setq aw-keys '(?a ?s ?d ?f ?g ?h ?j ?k ?l))
   :bind ("C-x o" . ace-window))
+
+;; (use-package beacon
+;;   :diminish beacon-mode
+;;   :config (progn
+;;             (beacon-mode 1)
+;;             (setq beacon-blink-delay 0.2
+;;                   beacon-blink-duration 0.2
+;;                   beacon-blink-when-point-moves 7
+;;                   beacon-blink-when-window-changes t
+;;                   beacon-blink-when-window-scrolls t
+;;                   beacon-color "brown"
+;;                   beacon-push-mark 5
+;;                   beacon-size 25)))
 
 (provide '01_basic_ui)
 ;;; 01_basic_ui.el ends here