+++ src/scame/packages/04_search.el
@@ -20,83 +20,98 @@
 ;;; Code:
 
 
-(use-package ido
-  ;;:init (ido-mode 1)
-  :config (progn
-            (setq ido-case-fold t)
-            (setq ido-everywhere t)
-            (setq ido-enable-prefix nil)
-            (setq ido-enable-flex-matching t)
-            (setq ido-create-new-buffer 'always)
-            (setq ido-max-prospects 10)
-            (add-to-list 'ido-ignore-files "\\.DS_Store")))
+
+(global-unset-key (kbd "C-x c"))
+(define-prefix-command 'scame-completion-map)
+(global-set-key (kbd "C-x c") 'scame-completion-map)
 
 
-(global-unset-key (kbd "C-x h"))
-(define-prefix-command 'scame-helm-map)
-(global-set-key (kbd "C-x h") 'scame-helm-map)
+(defun scame--setup-ido ()
+  "Setup Scame with IDO backend."
+  (use-package ido
+    ;;:init (ido-mode 1)
+    :config (progn
+              (setq ido-case-fold t)
+              (setq ido-everywhere t)
+              (setq ido-enable-prefix nil)
+              (setq ido-enable-flex-matching t)
+              (setq ido-create-new-buffer 'always)
+              (setq ido-max-prospects 10)
+              (add-to-list 'ido-ignore-files "\\.DS_Store"))))
+
+(defun scame--setup-helm ()
+  "Setup Scame with HELM backend."
+
+  (use-package helm
+    :bind (("C-x c F" . helm-find-files)
+           ("C-x c f" . helm-for-files)
+           ("C-x c r" . helm-recentf)
+           ("C-x c x" . helm-M-x)
+           ("C-x c y" . helm-show-kill-ring)
+           ("C-x c l" . helm-bookmarks)
+           ("C-x c a" . helm-apropos)
+           ("C-x c i" . helm-info-emacs)
+           ("C-x c b" . helm-buffers-list)))
+
+  (use-package helm-occur
+    :bind (("C-x c o" . helm-occur)))
+
+  (use-package helm-imenu
+    :bind (("C-x c i" . helm-imenu)))
+
+  (use-package helm-descbinds
+    :bind ("C-x c h" . helm-descbinds))
+
+  (use-package helm-swoop
+    :config (progn
+              (setq helm-multi-swoop-edit-save t)
+              (setq helm-swoop-split-direction 'split-window-vertically))
+    :bind (("C-x c w" . helm-swoop)
+           ("C-x c W" . helm-multi-swoop)))
+  )
 
 
-(use-package helm
-  ;; :init (progn
-  ;;         (helm-mode 1))
-  ;; :config (progn
-  ;;           (setq helm-ff-tramp-not-fancy nil)
-  ;;           (setq helm-ff-skip-boring-files t)
-  ;;           (add-to-list 'helm-boring-buffer-regexp-list "\\*tramp")
-  ;;           (add-to-list 'helm-boring-buffer-regexp-list "\\*epc")
-  ;;           (add-to-list 'helm-boring-buffer-regexp-list "\\.elc$")
-  ;;           (add-to-list 'helm-boring-buffer-regexp-list "\\.pyc$"))
-  :bind (("C-x h F" . helm-find-files)
-         ("C-x h f" . helm-for-files)
-         ("C-x h r" . helm-recentf)
-         ("C-x h x" . helm-M-x)
-         ("C-x h y" . helm-show-kill-ring)
-         ("C-x h l" . helm-bookmarks)
-         ("C-x h a" . helm-apropos)
-         ("C-x h i" . helm-info-emacs)
-	 ("C-x h b" . helm-buffers-list)))
+(defun scame--setup-ivy ()
+    "Setup Scame using IVY backend."
+  (use-package counsel
+    :bind (("C-x c F" . counsel-find-file)
+           ("C-x c v" . counsel-describe-variable)
+           ("C-x c f" . counsel-describe-function)
+           ("C-x c y" . counsel-info-lookup-symbol)
+           ("C-x c i" . counsel-imenu)))
 
-(use-package helm-occur
-  :bind (("C-x h o" . helm-occur)))
+  (use-package ivy
+    :diminish ivy-mode
+    :bind (("C-x c r" . ivy-recentf)
+           ("C-x c b" . ivy-switch-buffer)))
 
-(use-package helm-imenu
-  :bind (("C-x h i" . helm-imenu)))
+  (use-package swiper
+    :bind (("C-s" . swiper)
+           ("C-r" . swiper)))
+  (autoload 'ivy-read "ivy"))
 
-;; (use-package helm-proc)
 
-(use-package helm-descbinds
-  :bind ("C-x h h" . helm-descbinds))
+(defun scame--setup-search-framework (completion-method)
+  "Setup search frameworks using `COMPLETION-METHOD'."
+  (cond ((eql 'ido completion-method)
+         (progn
+           (scame--setup-ido)
+           (ido-mode 1)))
+        ((eql 'helm completion-method)
+         (progn
+           (scame--setup-helm)
+           (helm-mode 1)))
+        ((eql 'ivy completion-method)
+         (progn
+           (scame--setup-ivy)
+           (ivy-mode t)))
+        (t (progn
+             (scame--setup-ido)
+             (ido-mode 1)))))
 
-(use-package helm-swoop
-  :config (progn
-            (setq helm-multi-swoop-edit-save t)
-            (setq helm-swoop-split-direction 'split-window-vertically))
-  :bind (("C-x h w" . helm-swoop)
-         ("C-x h W" . helm-multi-swoop)))
 
-;; (use-package helm-dash
-;;   :config (progn
-;;             (setq helm-dash-min-length 1)
-;;             (setq helm-dash-browser-func 'eww)))
+(scame--setup-search-framework scame-completion-method)
 
-(use-package ivy
-  :diminish ivy-mode)
-
-(use-package swiper
-  :bind (("C-s" . swiper)
-         ("C-r" . swiper)))
-(autoload 'ivy-read "ivy")
-
-;; (use-package swiper-helm)
-
-(cond ((eql 'ido scame-completion-method)
-       (ido-mode 1))
-      ((eql 'helm scame-completion-method)
-       (helm-mode 1))
-      ((eql 'ivy scame-completion-method)
-       (ivy-mode t))
-      (t (ido-mode 1)))