+++ src/scame/packages/13_modeline.el
@@ -1,6 +1,6 @@
 ;;; 13_modeline.el --- Mode-line configuration
 
-;; Copyright (c) 2014, 2015 Nicolas Lamirault <nicolas.lamirault@gmail.com>
+;; Copyright (c) 2014, 2015, 2016 Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 ;; This program is free software: you can redistribute it and/or modify
 ;; it under the terms of the GNU General Public License as published by
@@ -69,6 +69,31 @@
 ;;    (global :when active)
 ;;    buffer-position
 ;;    hud))
+
+(use-package spaceline
+  :config (use-package spaceline-config
+            :config
+            (setq powerline-default-separator 'wave)
+            (spaceline-define-segment line-column
+                                      "The current line and column numbers."
+                                      "l:%l c:%2c")
+            ;;(spaceline-helm-mode)
+            (spaceline-toggle-battery-on)
+            (spaceline-toggle-minor-modes-on)
+            (spaceline-toggle-flycheck-info-off)
+            (spaceline-toggle-buffer-encoding-off)
+            (spaceline-toggle-buffer-encoding-abbrev-off)
+            (spaceline-toggle-buffer-size-off)
+            (spaceline-toggle-persp-name-on)
+            ;; (spaceline-define-segment time
+            ;;                           "The current time."
+            ;;                           (format-time-string "%H:%M"))
+            ;; (spaceline-define-segment date
+            ;;                           "The current date."
+            ;;                           (format-time-string "%h %d"))
+            ;; (spaceline-toggle-time-on)
+            (spaceline-emacs-theme 'date 'time)))
+
 
 
 (provide '13_modeline)