+++ src/scame/packages/34_go-lang.el
@@ -67,6 +67,18 @@
                         (lambda ()
                           (local-set-key (kbd "C-x g l") 'golint)))))
 
+  (use-package go-rename
+    :config (progn
+              (add-hook 'go-mode-hook
+                        (lambda ()
+                          (local-set-key (kbd "C-x g R") 'go-rename)))))
+
+  (use-package go-dlv
+    :config (progn
+              (add-hook 'go-mode-hook
+                        (lambda ()
+                          (local-set-key (kbd "C-x g d") 'dlv-current-func)))))
+
   (use-package go-direx
     ;; :defer scame-defer-package
     :config (progn
@@ -85,6 +97,25 @@
                       (lambda ()
                         (set (make-local-variable 'company-backends) '(company-go))
                         (company-mode))))
+
+  (define-key go-mode-map (kbd "C-x g h")
+    (defhydra hydra-go (:color blue)
+      "
+   ^Test^                ^Tools^
+  ╭─────────────────────────────────
+   _t_: test            _d_: godoc
+   _f_: file            _e_: errcheck
+   _p_: project         _l_: golint
+  ----------------------------------
+        "
+      ("q" nil "quit")
+      ("t" go-test-current-test "test")
+      ("f" go-test-current-file "file")
+      ("p" go-test-current-project "project")
+      ("d" godoc "godoc")
+      ("e" go-errcheck "errcheck")
+      ("l" golint "golint")
+      ))
 
 )