+++ src/scame/packages/44_rust.el
@@ -1,6 +1,6 @@
 ;;; 44_rust.el -- Rust configuration
 
-;; Copyright (C) 2014, 2015  Nicolas Lamirault <nicolas.lamirault@gmail.com>
+;; Copyright (C) 2014, 2015, 2016  Nicolas Lamirault <nicolas.lamirault@gmail.com>
 
 ;; This program is free software: you can redistribute it and/or modify
 ;; it under the terms of the GNU General Public License as published by
@@ -22,13 +22,19 @@
 (when scame-rust
 
   (use-package rust-mode
-    ;; :defer scame-defer-package
     :mode (("\\.rs\\'" . rust-mode))
     :init (add-hook 'rust-mode-hook 'flycheck-mode)
     :config (setq tab-width 4))
 
-  ;; (use-package cargo
-  ;;   :config (add-hook 'rust-mode-hook 'cargo-minor-mode))
+  (use-package cargo
+    :config (progn
+              (add-hook 'rust-mode-hook 'cargo-minor-mode)
+              (add-hook 'rust-mode-hook
+                        (lambda ()
+                          (local-set-key (kbd "C-x r t") 'cargo-test-current-test)
+                          (local-set-key (kbd "C-x r f") 'cargo-test-current-file)
+                          (local-set-key (kbd "C-x r p") 'cargo-test-current-project)
+                          ))))
 
   (use-package racer
     :init (progn
@@ -46,6 +52,8 @@
 
   (use-package company-racer
     :init (set (make-local-variable 'company-backends) '(company-racer)))
+
+
 
   )